backend/requirements.txt
alembic==1.13.1
annotated-types==0.7.0
anyio==4.4.0
async-timeout==4.0.3
bcrypt==4.1.3
certifi==2024.2.2
cffi==1.16.0
click==8.1.7
cryptography==42.0.7
dnspython==2.6.1
ecdsa==0.19.0
email_validator==2.1.1
exceptiongroup==1.2.1
fastapi==0.111.0
fastapi-cli==0.0.4
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
idna==3.7
iniconfig==2.0.0
Jinja2==3.1.4
Mako==1.3.5
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
orjson==3.10.3
packaging==24.0
passlib==1.7.4
pluggy==1.5.0
psycopg==3.1.19
psycopg-binary==3.1.19
pyasn1==0.6.0
pycparser==2.22
pydantic==2.7.2
pydantic-settings==2.2.1
pydantic_core==2.18.3
Pygments==2.18.0
pytest==8.2.1
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.9
PyYAML==6.0.1
redis==5.0.4
rich==13.7.1
rsa==4.9
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.30
starlette==0.37.2
tenacity==8.3.0
tomli==2.0.1
typer==0.12.3
typing_extensions==4.12.0
ujson==5.10.0
uvicorn==0.30.0
uvloop==0.19.0
watchfiles==0.22.0
websockets==12.0
